+++ NEWS.md
@@ -1,3 +1,7 @@
+# sagemaker.core 0.2.1:
+## Bug Fix:
+* Ensure deserializer classes correctly parse raw vectors as expected.
+
 # sagemaker.core 0.2.0:
 ## Feature:
 * Add SparseMatrixSerializer serialize class